Move the World Bank loader into the data layer
Landing in climate_risk/data puts it inside mypy's include list, which is what the kuznets stub override is for.
