+++ climate_risk/data_functions/__init__.py
@@ -3,6 +3,7 @@ from climate_risk.data.gpcc import load_gpcc_data
 from climate_risk.data.hadcrut import load_hadcrut_data
 from climate_risk.data.ipcc import process_ipcc_scenarios
 from climate_risk.data.ocean_heat import load_ocean_heat_data
+from climate_risk.data.world_bank import load_wb_data
 from climate_risk.data_functions.combine_data import load_all_data
 from climate_risk.data_functions.disaster_point_data import (
     load_disaster_point_data,
@@ -11,7 +12,6 @@ from climate_risk.data_functions.disaster_point_data import (
 from climate_risk.data_functions.emdat_processing import load_emdat_data
 from climate_risk.data_functions.rivers_data_loader import load_rivers_data
 from climate_risk.data_functions.shapefiles_data_loader import load_shapefile
-from climate_risk.data_functions.world_bank_data_loader import load_wb_data
 
 __all__ = [
     "load_all_data",
