+++ climate_risk/data_functions/combine_data.py
@@ -6,8 +6,8 @@ import pandas as pd
 from climate_risk.data.co2 import load_co2_data
 from climate_risk.data.gpcc import load_gpcc_data
 from climate_risk.data.ocean_heat import load_ocean_heat_data
+from climate_risk.data.world_bank import load_wb_data
 from climate_risk.data_functions.emdat_processing import DISASTER_TYPES, load_emdat_data
-from climate_risk.data_functions.world_bank_data_loader import load_wb_data
 
 
 def load_all_data(cache_dir: Path) -> dict[str, pd.DataFrame]:
