+++ pyproject.toml
@@ -164,7 +164,7 @@ strict_equality = true
 
 # None of these four ships stubs, and none exists on the index.
 [[tool.mypy.overrides]]
-module = ["pymc.*", "pytensor.*", "sklearn.*", "statsmodels.*"]
+module = ["kuznets.*", "pymc.*", "pytensor.*", "sklearn.*", "statsmodels.*"]
 ignore_missing_imports = true
 
 # Test functions carry no annotations, so their bodies need checking on their own terms.
